refactor(app): rename Paragraph component
