+++ packages/app/modules/editor/components/NodeEditorBlock.vue
@@ -58,6 +58,69 @@ onMounted(() => {
 onUnmounted(() => {
     window.removeEventListener('focus', setIsFocused)
 })
+
+// keybindings
+
+// const manager = useManger()
+// const focusList = useFocusList('[data-block] [contenteditable]')
+// const cursor = useCursorHelper()
+
+// const content = ref<HTMLElement>()
+
+// function onBackspace(e: KeyboardEvent) {
+//     if (!cursor.isCaretOnStart()) return
+
+//     e.preventDefault()
+
+//     focusList.focus(-1)
+
+//     cursor.setCaretOnEnd()
+
+//     manager.removeNode(model.value)
+// }
+
+// function onDeleteKeypress(e: KeyboardEvent) {
+//     const haveText = model.value.toText().trim().length > 0
+
+//     if (haveText) return
+
+//     e.preventDefault()
+
+//     manager.removeNode(model.value)
+// }
+
+// function addNewNode(e: KeyboardEvent) {
+//     e.preventDefault()
+
+//     const paragraph = new MarkdownNode({
+//         type: NodeType.Paragraph,
+//         tokens: [Token.breakLine()],
+//     })
+
+//     if (cursor.isCaretOnStart()) {
+//         manager.addNodeBefore(model.value, paragraph)
+//     }
+
+//     if (!cursor.isCaretOnStart()) {
+//         manager.addNodeAfter(model.value, paragraph)
+//     }
+
+//     setTimeout(() => {
+//         focusList.focus(1)
+//     }, 50)
+// }
+
+// onKeyStroke('Backspace', onBackspace, {
+//     target: content,
+// })
+
+// onKeyStroke('Delete', onDeleteKeypress, {
+//     target: content,
+// })
+
+// onKeyStroke('Enter', addNewNode, {
+//     target: content,
+// })
 </script>
 
 <template>
